block_copy_study/requirements.txt
numpy
matplotlib
seaborn
pandas
pupil-labs-realtime-api
mbientlab-metawear
scipy
